Added comments in script and created a Google CoLab notebook
Cleaned up the code. Moved the random seed to execute_simulation.py instead of having it in epileptor_model.py. Not sure if this is the right thing to do. May need to change this again...

Created a Google CoLab notebook that runs the experiment. This is incase the simulation is too slow on my laptop.