Epileptor changes, state space changes, and agent-observe change
Too many changes for a single commit...

I reduced the sampling period of the Epileptor to reduce the total simulation time. Also changed some of the other time constants pertaining to the epileptor including seizure frequency. Increaed the amount of noise in the system. Also changed the filters to extract states. Changed the probability of choosing a random action and the rate of decay.

Realized that the observe method for the Agent class wasn't doing what it was supposed to be doing. I removed the external "step" which was iterating at each sample point rather than the internal "step" which incremented at each observation.